requirements.txt
# requirements.txt
sentence-transformers
faiss-cpu
beautifulsoup4
nltk
google-generativeai
python-dotenv